+++ sentinel_test.go
@@ -129,6 +129,15 @@ func (s *sentinelStub) switchPrimary(newPrimAddr string, newSecAddrs ...string) 
 	}
 }
 
+type clientWrapAddr struct {
+	Client
+	addr net.Addr
+}
+
+func (c clientWrapAddr) Addr() net.Addr {
+	return c.addr
+}
+
 func TestSentinel(t *T) {
 	ctx := testCtx(t)
 	stub := newSentinelStub(
@@ -139,8 +148,9 @@ func TestSentinel(t *T) {
 
 	// our fake poolFn will always _actually_ connect to 127.0.0.1, we just
 	// don't tell anyone
-	poolFn := func(ctx context.Context, _ string, _ string) (Client, error) {
-		return NewPool(ctx, "tcp", "127.0.0.1:6379", 1)
+	poolFn := func(ctx context.Context, network, addr string) (Client, error) {
+		c, err := NewPool(ctx, "tcp", "127.0.0.1:6379", 1)
+		return clientWrapAddr{Client: c, addr: rawAddr{network, addr}}, err
 	}
 
 	scc, err := NewSentinel(
@@ -150,20 +160,23 @@ func TestSentinel(t *T) {
 	require.Nil(t, err)
 
 	assertState := func(primAddr string, secAddrs, sentAddrs []string) {
-		gotPrimAddr, gotSecAddrs := scc.Addrs()
-		assert.Equal(t, primAddr, gotPrimAddr)
-		assert.Len(t, gotSecAddrs, len(secAddrs))
-		for i := range secAddrs {
-			assert.Contains(t, gotSecAddrs, secAddrs[i])
+		clients, err := scc.Clients()
+		assert.NoError(t, err)
+		assert.Contains(t, clients, primAddr)
+		assert.Equal(t, primAddr, clients[primAddr].Primary.Addr().String())
+		assert.Len(t, clients[primAddr].Secondaries, len(secAddrs))
+		for _, secondary := range clients[primAddr].Secondaries {
+			assert.Contains(t, secAddrs, secondary.Addr().String())
 		}
 
-		gotSentAddrs := scc.SentinelAddrs()
+		gotSentAddrs, err := scc.SentinelAddrs()
+		assert.NoError(t, err)
 		assert.Len(t, gotSentAddrs, len(sentAddrs))
 		for i := range sentAddrs {
 			assert.Contains(t, gotSentAddrs, sentAddrs[i])
 		}
 
-		err := scc.proc.WithRLock(func() error {
+		err = scc.proc.WithRLock(func() error {
 			assert.Len(t, scc.sentinelAddrs, len(sentAddrs))
 			for i := range sentAddrs {
 				assert.Contains(t, scc.sentinelAddrs, sentAddrs[i])
@@ -228,202 +241,6 @@ func (ssp *stubSentinelPool) Close() error {
 	return nil
 }
 
-// this also tests that Clients get carried over during failover.
-func TestSentinelClientsAddrs(t *T) {
-	ctx := testCtx(t)
-
-	type testState struct {
-		primAddr              string
-		nilSecAddrs, secAddrs []string
-	}
-
-	secAddrsM := func(ts testState) map[string]bool {
-		m := map[string]bool{}
-		for _, addr := range ts.nilSecAddrs {
-			m[addr] = true
-		}
-		for _, addr := range ts.secAddrs {
-			m[addr] = true
-		}
-		return m
-	}
-
-	secAddrs := func(ts testState) []string {
-		m := secAddrsM(ts)
-		l := make([]string, 0, len(m))
-		for addr := range m {
-			l = append(l, addr)
-		}
-		return l
-	}
-
-	assertAddrs := func(ts testState, sc *Sentinel) {
-		gotPrimAddr, gotSecAddrs := sc.Addrs()
-		assert.Equal(t, ts.primAddr, gotPrimAddr)
-
-		expSecAddrs := secAddrsM(ts)
-		assert.Len(t, gotSecAddrs, len(expSecAddrs))
-		for addr := range expSecAddrs {
-			assert.Contains(t, gotSecAddrs, addr)
-		}
-	}
-
-	type testCase struct {
-		start, end testState
-		closed     []string
-	}
-
-	poolFn := func(ctx context.Context, network, addr string) (Client, error) {
-		return &stubSentinelPool{addr: addr}, nil
-	}
-
-	cases := []testCase{
-		{
-			start:  testState{primAddr: "A:0"},
-			end:    testState{primAddr: "B:0"},
-			closed: []string{"A:0"},
-		},
-		{
-			start: testState{primAddr: "A:0"},
-			end:   testState{primAddr: "B:0", secAddrs: []string{"A:0"}},
-		},
-		{
-			start: testState{
-				primAddr:    "A:0",
-				nilSecAddrs: []string{"B:0"},
-			},
-			end: testState{primAddr: "B:0", secAddrs: []string{"A:0"}},
-		},
-		{
-			start: testState{
-				primAddr:    "A:0",
-				nilSecAddrs: []string{"B:0", "C:0"},
-			},
-			end: testState{
-				primAddr:    "B:0",
-				nilSecAddrs: []string{"C:0"},
-			},
-			closed: []string{"A:0"},
-		},
-		{
-			start: testState{
-				primAddr:    "A:0",
-				nilSecAddrs: []string{"B:0"},
-				secAddrs:    []string{"C:0"},
-			},
-			end: testState{
-				primAddr: "B:0",
-				secAddrs: []string{"C:0"},
-			},
-			closed: []string{"A:0"},
-		},
-		{
-			start: testState{
-				primAddr:    "A:0",
-				nilSecAddrs: []string{"B:0"},
-				secAddrs:    []string{"C:0"},
-			},
-			end: testState{
-				primAddr:    "A:0",
-				nilSecAddrs: []string{"B:0"},
-			},
-			closed: []string{"C:0"},
-		},
-		{
-			start: testState{
-				primAddr:    "A:0",
-				nilSecAddrs: []string{"B:0"},
-				secAddrs:    []string{"C:0"},
-			},
-			end: testState{
-				primAddr: "A:0",
-				secAddrs: []string{"C:0"},
-			},
-		},
-	}
-
-	for _, tc := range cases {
-		stub := newSentinelStub(tc.start.primAddr, secAddrs(tc.start), []string{"127.0.0.1:26379"})
-
-		sc, err := NewSentinel(
-			ctx, "stub", stub.sentAddrs,
-			SentinelConnFunc(stub.newConn), SentinelPoolFunc(poolFn),
-		)
-		require.Nil(t, err)
-
-		// call Client on all secAddrs so Clients get created for them, double
-		// check that the clients were indeed created in clients map
-		for _, addr := range tc.start.secAddrs {
-			client, err := sc.Client(ctx, addr)
-			assert.Nil(t, err)
-			assert.NotNil(t, client)
-			assert.Equal(t, client, sc.clients[addr])
-		}
-
-		// collect all non-nil clients to check against return from Clients
-		// later
-		prevClients := map[string]Client{}
-		for addr, client := range sc.clients {
-			if client != nil {
-				prevClients[addr] = client
-			}
-		}
-
-		// collect all clients which are expected to be closed, so we can check
-		// their closed fields later
-		willClose := map[string]*stubSentinelPool{}
-		for _, addr := range tc.closed {
-			client := sc.clients[addr]
-			require.NotNil(t, client)
-			willClose[addr] = client.(*stubSentinelPool)
-		}
-
-		assertAddrs(tc.start, sc)
-
-		stub.switchPrimary(tc.end.primAddr, secAddrs(tc.end)...)
-		assert.Equal(t, "switch-master completed", <-sc.testEventCh)
-
-		assertAddrs(tc.end, sc)
-		for addr, ssp := range willClose {
-			assert.True(t, ssp.closed, "addr:%q not closed", addr)
-		}
-
-		// check returns from Client. If the addr was in prevClients the Client
-		// should stay the same from there.
-		assertClient := func(addr string) {
-			assert.Contains(t, sc.clients, addr)
-			client, err := sc.Client(ctx, addr)
-			assert.Nil(t, err)
-			if prevClient := prevClients[addr]; prevClient != nil {
-				assert.Equal(t, prevClient, client)
-			}
-		}
-		assertClient(tc.end.primAddr)
-		for _, secAddr := range tc.end.secAddrs {
-			assertClient(secAddr)
-		}
-
-		// test that, for nilSecAddrs, they are in the clients map but don't
-		// have a Client value. Then test that if Client is called with that
-		// addr a new Client is created in the clients map, and that same client
-		// is returned the next time Client is called.
-		for _, nilSecAddr := range tc.end.nilSecAddrs {
-			assert.Contains(t, sc.clients, nilSecAddr)
-			assert.Nil(t, sc.clients[nilSecAddr])
-
-			client, err := sc.Client(ctx, nilSecAddr)
-			assert.Nil(t, err)
-			assert.NotNil(t, client)
-			assert.Equal(t, client, sc.clients[nilSecAddr])
-
-			client2, err := sc.Client(ctx, nilSecAddr)
-			assert.Nil(t, err)
-			assert.Equal(t, client, client2)
-		}
-	}
-
-}
-
 func TestSentinelSecondaryRead(t *T) {
 	ctx := testCtx(t)
 	stub := newSentinelStub(
@@ -449,12 +266,22 @@ func TestSentinelSecondaryRead(t *T) {
 	require.Nil(t, err)
 
 	runTest := func(n int) {
-		primAddr, secAddrs := scc.Addrs()
+		clients, err := scc.Clients()
+		assert.NoError(t, err)
+		var primAddr string
+		for primAddr = range clients {
+			break
+		}
+
 		for i := 0; i < n; i++ {
 			var addr string
 			require.NoError(t, scc.DoSecondary(ctx, Cmd(&addr, "GIMME", "YOUR", "ADDRESS")))
-			assert.NotEqualf(t, scc.primAddr, addr, "command was sent to master at %s", primAddr)
-			assert.Containsf(t, secAddrs, addr, "returned address if not a secondary. expected one of %v, got %v", secAddrs, addr)
+			assert.NotEqualf(t, primAddr, addr, "command was sent to master at %s", primAddr)
+			var secAddrs []string
+			for _, secondary := range clients[primAddr].Secondaries {
+				secAddrs = append(secAddrs, secondary.Addr().String())
+			}
+			assert.Containsf(t, secAddrs, addr, "returned address is not a secondary or primary")
 		}
 	}